requirements.txt
asn1crypto==1.5.1
cffi==1.15.1
cryptography==2.3
idna==3.4
pycparser==2.21
PyJWT==1.6.4
six==1.16.0
sqlalchemy==1.4.44
jsonschema==4.17.3